Merge jaypipes/gdt-core and jaypipes/gdt repos
Merges the `github.com/jaypipes/gdt` and `github.com/jaypipes/gdt-core`
repos into a single `github.com/gdt-dev/gdt` repository.

In the process of merging these repositories, I reworked the awkward way
that plugins were stored in a context and then a Scenario had a ctx
field to grab the plugins from. This was mostly done to ease in testing
plugins, but it just made the calling interface awkward, so I now just
pull the registered plugins at the start of a Scenario's parse, which
means it's easier and less verbose for both porcelain callers like
`gdt.From()` and plumbing callers like `scenario.FromReader()`.

The exec plugin is now a real plugin and registers itself in the
`init()` of the main `gdt` package namespace.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>