feature: skip-if conditional scenario skip
Adds support for `skip-if` collection of evaluable conditions for a
scenario. If any of these conditions succeed, the test will be skipped.
This allows test authors to specify "pre-flight checks" before
attempting any of the actions in the scenario's tests.

For example, let's assume you have a `gdt-kube` scenario that looks like
this:

```yaml
tests:
 - kube.create: manifests/nginx-deployment.yaml
 - kube:
   get: deployments/nginx
   assert:
     matches:
       status:
         readyReplicas: 2
 - kube.delete: deployments/nginx
```

If you execute the above test and there is already an 'nginx'
deployment, the `kube.create` test will fail. To prevent the scenario
from proceeding with the tests if an 'nginx' deployment already exists,
you could add the following

```yaml
skip-if:
 - kube.get: deployments/nginx
tests:
 - kube.create: manifests/nginx-deployment.yaml
 - kube:
   get: deployments/nginx
   assert:
     matches:
       status:
         readyReplicas: 2
 - kube.delete: deployments/nginx
```

With the above, if an 'nginx' deployment exists already, the scenario
will skip all the tests.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>