nest `exec` plugin assertions under `assert` field
I'm aligning all of the plugin assertion fields with a name of `assert`.
This commit changes the first plugin, `exec` to nest the `exit_code`,
`err` and `out` assertion structs under an `assert` field object in the
Spec.

So, instead of this:

```yaml
tests:
 - exec: echo cat
   out:
     is: cat
```

you now do this:

```yaml
tests:
 - exec: echo cat
   assert:
     out:
       is: cat
```

I'm changing the `http` plugin's assertions block from `response` to
`assert` and the `kube` plugin's assertions block from `kube.assert` to
just `assert` in followup PRs in those repositories.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>