`assert.out.none` and `assert.err.none` to exec
Makes the assertion fields of the exec plugin's `assert` struct more
flexible and adds a new `assert.out.none` and `assert.err.none` field
that contains a string or strings that should *not* appear in either
stdout or stderr pipes from the exec command.

The `plugin/exec.PipeExpect` struct now contains three fields, each of
which are `types.FlexStrings` that can contain either a string or a list
of strings. There is now a custom YAML unmarshaler for `PipeExpect`
which allows a flexible naming convention to be used for specifying
assertions:

All of these examples are valid and represent human-readable assertions:

```yaml
tests:
  # Variants of contains
  - exec: ls -l
    assert:
      out:
        contains:
         - parse.go
  - exec: ls -l
    assert:
      out:
        is: parse.go
  - exec: ls -l
    assert:
      out:
        is:
         - parse.go
  - exec: ls -l
    assert:
      out:
        contains: parse.go
  - exec: ls -l
    assert:
      out:
        contains-all:
         - parse.go
  - exec: ls -l
    assert:
      out:
        all: parse.go
  # Variants of contains-any
  - exec: ls -l
    assert:
      out:
        any:
         - thisdoesnotexist
         - neitherdoesthisexist
         - parse.go
  - exec: ls -l
    assert:
      out:
        contains_one_of:
         - thisdoesnotexist
         - neitherdoesthisexist
         - parse.go
  - exec: ls -l
    assert:
      out:
        contains-any:
         - parse.go
  - exec: ls -l
    assert:
      out:
        any: parse.go
  # Variants of contains-none-of
  - exec: ls -l
    assert:
      out:
        none:
         - notexisting.go
  - exec: ls -l
    assert:
      out:
        none-of: notexisting.go
  - exec: ls -l
    assert:
      out:
        contains_none_of:
         - notexisting.go
  - exec: ls -l
    assert:
      out:
        none-of: notexisting.go
```

Signed-off-by: Jay Pipes <jaypipes@gmail.com>