exec: fix debug handler
For the exec plugin, we were writing debug output even when the user had
not called `gdtcontext.WithDebug()`. In addition, stdout and stderr
contents were only being written to debug writers when executed during
`on.fail.exec`. These two problems are fixed in this commit in addition
to correcting an erroneous `$LOCATION` in the README.md that needed to
be `$$LOCATION`.

Signed-off-by: Jay Pipes <jaypipes@gmail.com>