add custom retry configuration
Adds the ability to configure retry behaviour for any test spec, using
either a constant interval or an exponential backoff.

In doing so, reworks the framework to have all calls to
testing.T.Errorf() happen in the Scenario.Run() method, as well as have
that method handle all retries. gdt-kube will need to be refactored to
remove the retry behaviour implemented in that plugin.

Also, this finally fixes the way that failures are tested by using the
strategy outlined in https://github.com/golang/go/issues/39903.

Issue #26

Signed-off-by: Jay Pipes <jaypipes@gmail.com>