add variable save/restore for exec plugin
Adds support for defining variables in an exec plugin test spec and
referring to the variable in subsequent test specs.

A `gdt` test scenario is comprised of a list of test specs. These test specs
are executed in sequential order. If you want to have one test spec be able to
use some output or value calculated or asserted in a previous step, you can use
the `gdt` variable system.

Here's an test scenario that shows how to define variables in a test spec and
how to use those variables in later test specs.

file: `plugin/exec/testdata/var-save-restore.yaml`:

```yaml
name: var-save-restore
description: a scenario that tests variable save/restore across multiple test specs
tests:
  - exec: echo 42
    var-stdout: VAR_STDOUT

  - exec: echo $$VAR_STDOUT
    var-rc: VAR_RC
    assert:
      out:
        is: 42

  - exec: echo $$VAR_RC
    assert:
      out:
        is: 0

  - exec: echo 42
    assert:
      out:
        is: $$VAR_STDOUT
```

In the first test spec, we specify that we want to store the value of the
`stdout` stream in a variable called `VAR_STDOUT`:

```yaml
  - exec: echo 42
    var-stdout: VAR_STDOUT
```

In the second test spec, we refer to the `VAR_STDOUT` variable using the
double-dollar-sign notation in the `exec` field and also specify a `VAR_RC`
variable to contain the value of the return/exitcode from the executed
statement (`echo 42`):

```yaml
  - exec: echo $$VAR_STDOUT
    var-rc: VAR_RC
    assert:
      out:
        is: 42
```

> **NOTE**: We use the double-dollar-sign notation because by default, `gdt`
> replaces all single-dollar-sign notations with environment variables *BEFORE*
> executing the test specs in a test scenario. Using the double-dollar-sign
> notation means that environment variable substitution does not impact the
> referencing of `gdt` variables referenced in a test spec.

In the third test spec, we simply echo out the value of that `VAR_RC` variable
and assert that the stdout stream contains the string "0" (since `echo 42`
returns 0.):

```yaml
  - exec: echo $$VAR_RC
    assert:
      out:
        is: 0
```

Finally, in the fourth step, we demonstrate that we can refer to the
`VAR_STDOUT` variable defined in the very first test spec from the
`assert.out.is` field. This shows the flexibility of the `gdt` variable system.
You can define variables using a simple declarative syntax and then refer to
the value of those variables using the double-dollar-sign notation in any
subsequent test spec.

Issue #51

Signed-off-by: Jay Pipes <jaypipes@gmail.com>